+++ ui/components/subscription/TeamMembers.tsx
@@ -1,5 +1,7 @@
 import { useNFT } from '@thirdweb-dev/react'
 import { useEffect, useState } from 'react'
+import useHatNames from '@/lib/hats/useHatNames'
+import useUniqueHatWearers from '@/lib/hats/useUniqueHatWearers'
 import { useHandleRead } from '@/lib/thirdweb/hooks'
 import ProfileCard from '../layout/ProfileCard'
 
@@ -28,31 +30,7 @@ function TeamMember({
   hatsContract,
 }: TeamMemberProps) {
   //Hats Data
-  const [hatNames, setHatNames] = useState<string[]>([])
-
-  useEffect(() => {
-    async function getAllHatNames() {
-      try {
-        const hatNamesPromises = hatIds.map(async (hatId: string) => {
-          const hat = await hatsContract.call('viewHat', [hatId])
-          const detailsIpfsHash = hat.details.split('ipfs://')[1]
-          const hatDetailsRes = await fetch(
-            `https://ipfs.io/ipfs/${detailsIpfsHash}`
-          )
-          const { data: hatDetails } = await hatDetailsRes.json()
-          return hatDetails.name
-        })
-
-        const names = await Promise.all(hatNamesPromises)
-        setHatNames(names)
-      } catch (err) {
-        console.error('Failed to fetch hat names:', err)
-      }
-    }
-    if (hatIds.length > 0 && hatsContract) {
-      getAllHatNames()
-    }
-  }, [hatIds, hatsContract])
+  const hatNames = useHatNames(hatsContract, hatIds)
 
   //Citizen Data
   const { data: ownedToken } = useHandleRead(citizenContract, 'getOwnedToken', [
@@ -108,7 +86,7 @@ function TeamMember({
         subheader={
           <div className="flex flex-col h-[50px] overflow-auto">
             {hatNames?.map((hatName: any, i: number) => (
-              <p key={`${address}-hat-name-${i}`}>{hatName}</p>
+              <p key={`${address}-hat-name-${i}`}>{hatName.name}</p>
             ))}
           </div>
         }
@@ -123,32 +101,7 @@ export default function TeamMembers({
   citizenConract,
   hats,
 }: TeamMembersProps) {
-  const [wearers, setWearers] = useState<any>()
-
-  useEffect(() => {
-    async function hatsToUniqueWearers() {
-      const uniqueWearers: Wearer[] = []
-      hats.forEach((hat) => {
-        hat.wearers.forEach((w: any) => {
-          const existingWearer = uniqueWearers.find((u) => u.address === w.id)
-          if (existingWearer) {
-            existingWearer.hatIds.push(hat.id)
-          } else {
-            uniqueWearers.push({
-              address: w.id,
-              hatIds: [hat.id],
-            })
-          }
-        })
-      })
-      setWearers(uniqueWearers)
-    }
-
-    if (hats) {
-      hatsToUniqueWearers()
-    }
-  }, [hats])
-
+  const wearers = useUniqueHatWearers(hats)
   return (
     <>
       {wearers?.[0].address &&
